refactor(tui): extract s4 union (attachments_mixin w1b + billing_wire_mixin w1a) from server.py
