+++ tui_gateway/server.py
@@ -8435,149 +8435,6 @@ def _serialize_billing_state(state) -> dict:
     }
 
 
-def _usage_payload(state) -> dict:
-    """Best-effort shared usage model for the /topup + /subscription overlay bars.
-
-    Only fetched when logged in; fail-open to {available:false} so the overview
-    still renders if the account-info path is down.
-    """
-    if not getattr(state, "logged_in", False):
-        return {"available": False}
-    try:
-        from agent.billing_usage import build_usage_model
-
-        return _serialize_usage_model(build_usage_model())
-    except Exception:
-        return {"available": False}
-
-
-def _serialize_usage_bar(bar) -> Optional[dict]:
-    """Serialize a UsageBar (dollar magnitudes → display strings + fractions)."""
-    if bar is None:
-        return None
-    from agent.billing_usage import _fmt_usd
-
-    return {
-        "kind": bar.kind,
-        "remaining_display": _fmt_usd(bar.remaining_usd),
-        "total_display": _fmt_usd(bar.total_usd),
-        "spent_display": _fmt_usd(bar.spent_usd),
-        "pct_used": bar.pct_used,
-        "fill_fraction": bar.fill_fraction,
-    }
-
-
-def _serialize_usage_model(model) -> dict:
-    """Serialize a UsageModel for the wire — the shared two-bar dollar view.
-
-    Dollars-only (no 'credits'); fail-open shape mirrors the other billing RPCs
-    ({ok, available:false} when logged out / unreachable).
-    """
-    from agent.billing_usage import _fmt_usd, format_renews
-
-    if model is None or not getattr(model, "available", False):
-        return {"ok": True, "available": False}
-
-    return {
-        "ok": True,
-        "available": True,
-        "status": model.status,
-        "plan_name": model.plan_name,
-        "renews_at": model.renews_at,
-        "renews_display": getattr(model, "renews_display", None) or format_renews(model.renews_at),
-        "subscription_remaining_display": (
-            None if model.subscription_remaining_usd is None else _fmt_usd(model.subscription_remaining_usd)
-        ),
-        "topup_remaining_display": (
-            None if model.topup_remaining_usd is None else _fmt_usd(model.topup_remaining_usd)
-        ),
-        "total_spendable_display": (
-            None if model.total_spendable_usd is None else _fmt_usd(model.total_spendable_usd)
-        ),
-        "has_topup": model.has_topup,
-        "plan_bar": _serialize_usage_bar(model.plan_bar),
-        "topup_bar": _serialize_usage_bar(model.topup_bar),
-    }
-
-
-def _serialize_subscription_state(state) -> dict:
-    """Serialize a SubscriptionState for the wire (Decimals → strings)."""
-    from agent.billing_usage import format_renews
-    from agent.billing_view import format_money
-
-    def _s(value):
-        return None if value is None else str(value)
-
-    current = None
-    if state.current is not None:
-        c = state.current
-        current = {
-            "tier_id": c.tier_id,
-            "tier_name": c.tier_name,
-            "monthly_credits": _s(c.monthly_credits),
-            "credits_remaining": _s(c.credits_remaining),
-            "cycle_ends_at": c.cycle_ends_at,
-            "pending_downgrade_tier_name": c.pending_downgrade_tier_name,
-            "pending_downgrade_at": c.pending_downgrade_at,
-            "pending_downgrade_display": format_renews(c.pending_downgrade_at),
-            "cancel_at_period_end": c.cancel_at_period_end,
-            "cancellation_effective_at": c.cancellation_effective_at,
-            "cancellation_effective_display": format_renews(c.cancellation_effective_at),
-        }
-    # Selectable catalog for the in-terminal tier picker; price is pre-formatted
-    # ($X / $X.YY) so the TUI renders it directly.
-    tiers = [
-        {
-            "tier_id": t.tier_id,
-            "name": t.name,
-            "tier_order": t.tier_order,
-            "dollars_per_month_display": format_money(t.dollars_per_month),
-            "monthly_credits": _s(t.monthly_credits),
-            "is_current": t.is_current,
-            "is_enabled": t.is_enabled,
-        }
-        for t in state.tiers
-    ]
-    return {
-        "ok": True,
-        "logged_in": state.logged_in,
-        "is_admin": state.is_admin,
-        "can_change_plan": state.can_change_plan,
-        "org_name": state.org_name,
-        "org_id": state.org_id,
-        "role": state.role,
-        "context": state.context,
-        "current": current,
-        "tiers": tiers,
-        "portal_url": state.portal_url,
-        "error": state.error,
-        # Shared dollar usage model (two-bar view) embedded so /subscription
-        # renders the same bars as /usage from its single fetch. Built from the
-        # separate account-info path (the only source with top-up dollars);
-        # fail-open → {available:false}. Computed lazily so a logged-out state
-        # adds no cost.
-        "usage": _usage_payload(state),
-    }
-
-
-def _serialize_subscription_preview(p) -> dict:
-    """Serialize a SubscriptionChangePreview for the wire (Decimal → string)."""
-    return {
-        "ok": True,
-        "effect": p.effect,
-        "reason": p.reason,
-        "current_tier_id": p.current_tier_id,
-        "current_tier_name": p.current_tier_name,
-        "target_tier_id": p.target_tier_id,
-        "target_tier_name": p.target_tier_name,
-        "monthly_credits_delta": (
-            None if p.monthly_credits_delta is None else str(p.monthly_credits_delta)
-        ),
-        "amount_due_now_cents": p.amount_due_now_cents,
-        "effective_at": p.effective_at,
-    }
-
-
 # ── Delegation: subagent tree observability + controls ───────────────
 # Powers the TUI's /agents overlay (see ui-tui/src/components/agentsOverlay).
 # The registry lives in tools/delegate_tool — these handlers are thin
@@ -10205,241 +10062,7 @@ _IMAGE_MAGIC: tuple[tuple[bytes, str], ...] = (
 )
 
 
-def _decode_attach_base64(raw: str, *, mime_prefix: str) -> bytes | None:
-    """Decode a base64 (optionally data-URL-wrapped) payload.
-
-    Accepts ``data:<mime_prefix>...;base64,<b64>`` plus embedded whitespace.
-    Returns the decoded bytes, or ``None`` when the input isn't valid base64.
-    """
-    import base64 as _base64
-    import re as _re
-
-    cleaned = raw.strip()
-    m = _re.match(
-        rf"^data:{_re.escape(mime_prefix)}[a-zA-Z0-9.+-]*;base64,(.*)$",
-        cleaned,
-        _re.DOTALL,
-    )
-    if m:
-        cleaned = m.group(1)
-    cleaned = _re.sub(r"\s+", "", cleaned)
-    try:
-        return _base64.b64decode(cleaned, validate=True)
-    except Exception:
-        return None
-
-
-def _sniff_image_ext(img_bytes: bytes, filename: str = "") -> str:
-    """Resolve an image extension from a filename hint, else magic bytes.
-
-    Falls back to ``.png``. WebP needs the RIFF/WEBP container check, handled
-    before the generic table.
-    """
-    if filename:
-        suffix = Path(filename).suffix.lower()
-        if suffix:
-            return suffix
-    head = img_bytes[:16]
-    if head.startswith(b"RIFF") and head[8:12] == b"WEBP":
-        return ".webp"
-    for sig, ext in _IMAGE_MAGIC:
-        if head.startswith(sig):
-            return ext
-    return ".png"
-
-
-def _allowed_image_extensions() -> frozenset[str]:
-    try:
-        from cli import _IMAGE_EXTENSIONS
-
-        return frozenset(_IMAGE_EXTENSIONS)
-    except Exception:
-        return frozenset({".png", ".jpg", ".jpeg", ".gif", ".webp", ".bmp"})
-
-
-def _session_images_dir(session: dict) -> Path:
-    """Resolve the uploads ``images/`` dir against the session's effective home.
-
-    Attach RPCs (``image.attach_bytes``, ``clipboard.paste``, ``pdf.attach``)
-    run BEFORE ``prompt.submit`` installs the session's profile HERMES_HOME
-    override, so ``get_hermes_home()`` here would return the gateway's launch
-    home. In a multi-profile / root-gateway deployment that writes the upload to
-    the launch home's ``images/`` while the sandbox mount and the vision host-
-    read allowlist both resolve the *session profile's* ``images/`` at run time
-    — so the file the agent tries to read is never the file we wrote (#69575).
-
-    Anchor the write on the session's stored ``profile_home`` when present
-    (matching the mount/read scope), else fall back to the launch home. Keeps
-    per-profile isolation: a profile's uploads stay under that profile's home.
-    """
-    profile_home = session.get("profile_home")
-    base = Path(profile_home) if profile_home else _hermes_home
-    return base / "images"
-
-
-def _queue_attached_image(session: dict, img_bytes: bytes, ext: str, *, prefix: str) -> Path:
-    """Write image bytes into the gateway's images dir and queue them.
-
-    Mirrors what ``image.attach`` does for a local path: appends to
-    ``session["attached_images"]`` so the next ``prompt.submit`` picks it up via
-    the existing native-image-attach pipeline. Returns the written path.
-    """
-    session["image_counter"] = session.get("image_counter", 0) + 1
-    img_dir = _session_images_dir(session)
-    img_dir.mkdir(parents=True, exist_ok=True)
-    ts = datetime.now().strftime("%Y%m%d_%H%M%S")
-    img_path = img_dir / f"{prefix}_{ts}_{session['image_counter']}{ext}"
-    try:
-        img_path.write_bytes(img_bytes)
-    except Exception:
-        session["image_counter"] = max(0, session["image_counter"] - 1)
-        raise
-    session.setdefault("attached_images", []).append(str(img_path))
-    return img_path
-
-
 _ATTACHMENT_REF_NEEDS_QUOTING_RE = None
-
-
-def _format_ref_value(value: str) -> str:
-    """Quote a context-ref value when it contains whitespace or bracket chars.
-
-    Mirrors the desktop ``formatRefValue`` so the staged ``@file:`` ref round-trips
-    through ``agent.context_references`` cleanly.
-    """
-    import re as _re
-
-    global _ATTACHMENT_REF_NEEDS_QUOTING_RE
-    if _ATTACHMENT_REF_NEEDS_QUOTING_RE is None:
-        _ATTACHMENT_REF_NEEDS_QUOTING_RE = _re.compile(r"""[\s()\[\]{}<>"'`]""")
-    if not value or not _ATTACHMENT_REF_NEEDS_QUOTING_RE.search(value):
-        return value
-    if "`" not in value:
-        return f"`{value}`"
-    if '"' not in value:
-        return f'"{value}"'
-    if "'" not in value:
-        return f"'{value}'"
-    return value
-
-
-def _attachment_ref_path(session: dict, target: Path) -> str:
-    """Workspace-relative path for an attachment, or the absolute path if outside."""
-    workspace = Path(_session_cwd(session)).resolve()
-    try:
-        rel = target.resolve().relative_to(workspace)
-        return str(rel).replace(os.sep, "/")
-    except ValueError:
-        return str(target.resolve())
-
-
-def _desktop_attachment_dir(session: dict) -> Path:
-    root = Path(_session_cwd(session)).resolve() / ".hermes" / "desktop-attachments"
-    root.mkdir(parents=True, exist_ok=True)
-    return root
-
-
-def _sanitize_attachment_name(name: str) -> str:
-    import re as _re
-
-    candidate = Path(str(name or "").strip()).name
-    candidate = _re.sub(r"[\x00-\x1f]+", "_", candidate)
-    candidate = candidate.strip().strip(".")
-    return candidate or "attachment"
-
-
-def _unique_attachment_path(root: Path, filename: str) -> Path:
-    candidate = root / filename
-    if not candidate.exists():
-        return candidate
-    stem = Path(filename).stem or "attachment"
-    suffix = Path(filename).suffix
-    counter = 2
-    while True:
-        next_candidate = root / f"{stem}-{counter}{suffix}"
-        if not next_candidate.exists():
-            return next_candidate
-        counter += 1
-
-
-def _resolve_gateway_attachment_path(raw: str) -> Path | None:
-    """Resolve a raw path token to a gateway-visible file, or None."""
-    if not raw:
-        return None
-    try:
-        from cli import _detect_file_drop, _resolve_attachment_path, _split_path_input
-    except Exception:
-        return None
-
-    dropped = _detect_file_drop(raw)
-    if dropped:
-        return Path(dropped["path"]).resolve()
-    path_token, _remainder = _split_path_input(raw)
-    resolved = _resolve_attachment_path(path_token)
-    return Path(resolved).resolve() if resolved is not None else None
-
-
-def _decode_attachment_data_url(data_url: str) -> bytes:
-    """Decode a ``data:<any-mime>;base64,<b64>`` payload to bytes.
-
-    Unlike ``_decode_attach_base64`` (image-mime-specific), this accepts any
-    media type — text/csv, application/pdf, etc. — so non-image file uploads
-    round-trip. Also tolerates a bare base64 string with no data-URL prefix.
-    """
-    import base64 as _base64
-    import binascii as _binascii
-    import re as _re
-
-    cleaned = (data_url or "").strip()
-    m = _re.match(r"^data:[^;,]*(?:;[^;,=]+=[^;,]+)*;base64,(.*)$", cleaned, _re.DOTALL | _re.I)
-    if m:
-        cleaned = m.group(1)
-    cleaned = _re.sub(r"\s+", "", cleaned)
-    try:
-        return _base64.b64decode(cleaned, validate=True)
-    except (ValueError, _binascii.Error) as exc:
-        raise ValueError("invalid data_url payload") from exc
-
-
-def _stage_session_file_attachment(
-    session: dict,
-    *,
-    raw_path: str,
-    data_url: str,
-    name: str,
-) -> tuple[Path, bool]:
-    """Make a desktop file attachment available to the remote gateway agent.
-
-    Three cases:
-      1. The path resolves to a file already INSIDE the session workspace — use
-         it as-is (no copy, ``uploaded=False``).
-      2. The path resolves to a gateway-visible file OUTSIDE the workspace — copy
-         it into ``.hermes/desktop-attachments/`` so the ``@file:`` ref resolves.
-      3. The path doesn't exist on the gateway (the common remote case: it's a
-         path on the CLIENT's disk) — decode the uploaded ``data_url`` bytes and
-         write them into ``.hermes/desktop-attachments/``.
-
-    Returns ``(stored_path, uploaded)``.
-    """
-    workspace = Path(_session_cwd(session)).resolve()
-    resolved = _resolve_gateway_attachment_path(raw_path)
-    if resolved is not None:
-        try:
-            resolved.relative_to(workspace)
-            return resolved, False
-        except ValueError:
-            payload = resolved.read_bytes()
-            filename = resolved.name
-    else:
-        if not data_url:
-            raise ValueError("file not found on gateway and no data_url provided")
-        payload = _decode_attachment_data_url(data_url)
-        filename = _sanitize_attachment_name(name or Path(str(raw_path or "")).name)
-
-    upload_dir = _desktop_attachment_dir(session)
-    target = _unique_attachment_path(upload_dir, _sanitize_attachment_name(filename))
-    target.write_bytes(payload)
-    return target.resolve(), True
 
 
 # ── Methods: respond ─────────────────────────────────────────────────
@@ -13887,6 +13510,8 @@ def _browser_disconnect(rid) -> dict:
 # Imported at the end of this module so every global the handlers close
 # over already exists; register() rebinds them onto this namespace.
 from . import (  # noqa: E402
+    attachments_mixin as _attachments_mixin,
+    billing_wire_mixin as _billing_wire_mixin,
     methods_complete as _methods_complete,
     methods_config as _methods_config,
     methods_prompt as _methods_prompt,
@@ -13895,6 +13520,8 @@ from . import (  # noqa: E402
 )
 
 for _m in (
+    _attachments_mixin,
+    _billing_wire_mixin,
     _methods_session,
     _methods_prompt,
     _methods_config,
